requirements.txt
dtaidistance
dtw-python
faiss-gpu
h5py
hdf5plugin
matplotlib
numpy
pandas
Pillow
python-dateutil
pytorch-metric-learning
PyYAML
scikit-learn
scipy
tensorboardX
threadpoolctl
torch
torchaudio
torchvision
tqdm
typing
